+++ crates/partially/tests/derive/basic.rs
@@ -8,6 +8,33 @@ struct Data {
 }
 
 #[test]
+fn from_full_maps_all_some() {
+    let full = Data {
+        a: "a".to_string(),
+        b: "b".to_string(),
+    };
+    let partial: PartialData = full.into();
+    assert_eq!(partial.a, Some("a".to_string()));
+    assert_eq!(partial.b, Some("b".to_string()));
+}
+
+#[test]
+fn skip_from_full() {
+    #[derive(Partial)]
+    #[partially(skip_from_full)]
+    #[allow(unused)]
+    struct Struct {
+        a: String,
+    }
+    // NB: Without `skip_from_full` this would complain about a duplicate impl
+    impl From<Struct> for PartialStruct {
+        fn from(_: Struct) -> Self {
+            unimplemented!()
+        }
+    }
+}
+
+#[test]
 fn basic_apply_some() {
     let empty_partial = PartialData::default();
     let full_partial = PartialData {